added more helper functions, removed unused files and imports
